Summary
Added the speck files of the hwc and the labels of the hwc.
Added the script to create does speck and label files using the already created speck files.
Updated the hwcAssetCreation.py the GUI path was change.
Added the habitable world catalog asset. So you can see from the distance the exoplanets from the hwc.